Move and rename type
